Components: Generated the icav2 copy batch utility as a component
Removed the icav2 copy batch utility stack
