+++ config/constants.ts
@@ -109,16 +109,6 @@ export const icav2AccessTokenSecretName: Record<AppStage, string> = {
   [AppStage.PROD]: 'ICAv2JWTKey-umccr-prod-service-prod', // pragma: allowlist secret
 };
 
-/*
-Resources generated by the ICAv2 Copy Batch utility stack
- */
-export const icav2CopyBatchSSMRoot = '/orcabus/icav2_copy_batch_utility';
-export const icav2CopyBatchSSMBatchArn = path.join(icav2CopyBatchSSMRoot, 'batch_sfn_arn');
-export const icav2CopyBatchSSMBatchName = path.join(icav2CopyBatchSSMRoot, 'batch_sfn_name');
-export const icav2CopyBatchSSMSingleArn = path.join(icav2CopyBatchSSMRoot, 'single_sfn_arn');
-export const icav2CopyBatchSSMSingleName = path.join(icav2CopyBatchSSMRoot, 'single_sfn_name');
-export const icav2CopyBatchUtilityName = 'icav2_copy_batch_utility_sfn';
-export const icav2CopySingleUtilityName = 'icav2_copy_single_utility_sfn';
 
 /*
 Resources generated by the BSSH Fastq Copy Manager
