+++ lib/workload/components/icav2-copy-files-batch/step_functions_templates/copy_files_batch_sfn.asl.json
