+++ lib/workload/components/icav2-copy-files/check_or_launch_job_lambda_py/requirements.txt
@@ -1,0 +1,2 @@
+wrapica>=1.0.8
+boto3>=1.34
