+++ lib/workload/components/icav2-copy-files/step_functions_templates/copy_files_sfn.asl.json
